Doing a first pass on moving sim code to pdr-backend/simulation/ so that it's shared between predictoor and trader
