+++ pdr_backend/predictoor/approach3/predictoor_agent3.py
@@ -2,14 +2,14 @@ import os
 from typing import Tuple
 
 from enforce_typing import enforce_types
-from pdr_backend.predictoor.approach3.data_factory import DataFactory
-from pdr_backend.predictoor.approach3.model_factory import ModelFactory
-from pdr_backend.predictoor.approach3.model_ss import ModelSS
+from pdr_backend.predictoor.simulation.data_factory import DataFactory
+from pdr_backend.predictoor.simulation.model_factory import ModelFactory
+from pdr_backend.predictoor.simulation.model_ss import ModelSS
 
 from pdr_backend.predictoor.base_predictoor_agent import BasePredictoorAgent
 from pdr_backend.predictoor.approach3.predictoor_config3 import PredictoorConfig3
-from pdr_backend.predictoor.approach3.timeutil import timestr_to_ut
-from pdr_backend.predictoor.approach3.data_ss import DataSS
+from pdr_backend.predictoor.simulation.timeutil import timestr_to_ut
+from pdr_backend.predictoor.simulation.data_ss import DataSS
 
 
 @enforce_types
