+++ pdr_backend/predictoor/approach3/predictoor_config3.py
@@ -1,5 +1,5 @@
 from enforce_typing import enforce_types
-from pdr_backend.predictoor.approach3.timeutil import timestr_to_ut
+from pdr_backend.predictoor.simulation.timeutil import timestr_to_ut
 
 from pdr_backend.predictoor.base_predictoor_config import BasePredictoorConfig
