+++ pdr_backend/predictoor/simulation/test/conftest.py
@@ -1,0 +1,1 @@
+from pdr_backend.conftest_ganache import *  # pylint: disable=wildcard-import
