+++ pdr_backend/predictoor/simulation/test/test_timeblock.py
@@ -1,6 +1,6 @@
 from enforce_typing import enforce_types
 
-from pdr_backend.predictoor.approach3 import timeblock
+from pdr_backend.predictoor.simulation import timeblock
 
 
 @enforce_types
