+++ pdr_backend/predictoor/simulation/test/test_trade_engine.py
@@ -2,11 +2,11 @@ import os
 
 from enforce_typing import enforce_types
 
-from pdr_backend.predictoor.approach3.data_ss import DataSS
-from pdr_backend.predictoor.approach3.model_ss import ModelSS
-from pdr_backend.predictoor.approach3.timeutil import timestr_to_ut
-from pdr_backend.predictoor.approach3.tradeutil import TradeParams, TradeSS
-from pdr_backend.predictoor.approach3.trade_engine import TradeEngine
+from pdr_backend.predictoor.simulation.data_ss import DataSS
+from pdr_backend.predictoor.simulation.model_ss import ModelSS
+from pdr_backend.predictoor.simulation.timeutil import timestr_to_ut
+from pdr_backend.predictoor.simulation.tradeutil import TradeParams, TradeSS
+from pdr_backend.predictoor.simulation.trade_engine import TradeEngine
 
 
 @enforce_types
