+++ pdr_backend/predictoor/simulation/tradeutil.py
